delete_other_windowsしたときのoutput windowの位置を修正

+++ lib/textbringer/commands/windows.rb
@@ -48,6 +48,7 @@ module Textbringer
 
     define_command(:delete_output_window,
                    doc: "Delete output window.") do
+      raise EditorError, "Output window not exist" if Window.output.nil?
       Window.delete_window(Window.output)
       Window.output = nil
     end
